Using ModLib's mechanism for loading settings
Workaround for mods using FileDatabase with new ModLib
Fix for ModLib.Patches ModuleFolderName
SettingsProperty classes were moved to Models
Added IPropertyDefinitionWithEditableMinMax
Removed ScreenBase and MissionView references in MCM
MCM.UI should be loaded after SandBoxSubModule if placed on top
Mission leaving fix
MBO Min/Max values fix

+++ MCM.Implementation.MBO/Settings/Definitions/MBOPropertyDefinitionFloatingIntegerWrapper.cs
@@ -13,8 +13,8 @@ namespace MCM.Implementation.MBO.Settings.Definitions
 
         public MBOPropertyDefinitionFloatingIntegerWrapper(object @object) : base(@object)
         {
-            MinValue = @object.GetType().GetProperty(nameof(MinValue))?.GetValue(@object) as decimal? ?? 0;
-            MaxValue = @object.GetType().GetProperty(nameof(MaxValue))?.GetValue(@object) as decimal? ?? 0;
+            MinValue = (decimal) (@object.GetType().GetProperty(nameof(MinValue))?.GetValue(@object) as float? ?? 0);
+            MaxValue = (decimal) (@object.GetType().GetProperty(nameof(MaxValue))?.GetValue(@object) as float? ?? 0);
             ValueFormat = @object.GetType().GetProperty(nameof(ValueFormat))?.GetValue(@object) as string ?? "0.00";
         }
     }
